# agent-backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth": {
      "get": {
        "operationId": "AuthController_authenticateUser",
        "summary": "Authenticate User",
        "description": "Validates the current session/token and returns user details if authenticated. This endpoint is commonly used to check whether the user is logged in.",
        "tags": ["Auth"],
        "responses": {
          "200": {
            "description": "Successful authentication with user details.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "SUCCESS",
                      "default": "SUCCESS",
                      "description": "Indicates the result of the request"
                    },
                    "data": {
                      "type": "object",
                      "description": "Authenticated user information",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "user-id-placeholder"
                        },
                        "name": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "email": {
                          "type": "string",
                          "example": "user@example.com"
                        },
                        "isEmailVerified": {
                          "type": "boolean",
                          "example": true
                        },
                        "account": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "example": "account-id-placeholder"
                            },
                            "name": {
                              "type": "string",
                              "example": "Example Company"
                            }
                          }
                        },
                        "role": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "example": "role-id-placeholder"
                            },
                            "code": {
                              "type": "string",
                              "enum": ["superadmin", "admin", "user", "guest"],
                              "example": "admin",
                              "description": "Defines the role code of the authenticated user."
                            },
                            "value": {
                              "type": "string",
                              "enum": ["Admin", "Super Admin", "User", "Guest"],
                              "example": "Admin",
                              "description": "Human-readable role name for the authenticated user."
                            }
                          }
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "email",
                        "isEmailVerified",
                        "account",
                        "role"
                      ]
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2023-01-01T00:00:00.000Z",
                      "default": "2023-01-01T00:00:00.000Z"
                    }
                  },
                  "required": ["status", "data", "timestamp"]
                },
                "example": {
                  "status": "SUCCESS",
                  "data": {
                    "id": "user-id-placeholder",
                    "name": "John Doe",
                    "email": "user@example.com",
                    "isEmailVerified": true,
                    "account": {
                      "id": "account-id-placeholder",
                      "name": "Example Company"
                    },
                    "role": {
                      "id": "role-id-placeholder",
                      "code": "admin",
                      "value": "Admin"
                    }
                  },
                  "timestamp": "2023-01-01T00:00:00.000Z"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or token is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "FAILURE"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid token or session expired."
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "INVALID_TOKEN"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2023-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/agent-backend/auth"
                    }
                  },
                  "required": ["status", "errorCode", "timestamp", "path"]
                },
                "example": {
                  "status": "FAILURE",
                  "message": "Invalid token or session expired.",
                  "errorCode": "INVALID_TOKEN",
                  "timestamp": "2023-01-01T00:00:00.000Z",
                  "path": "/agent-backend/auth"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_loginUser",
        "summary": "User Login",
        "description": "Authenticates a user using their email and password credentials.",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              },
              "example": {
                "email": "user@example.com",
                "password": "yourStrongPassword123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns user profile and authentication status upon successful login.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "data", "timestamp"],
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "SUCCESS",
                      "default": "SUCCESS",
                      "description": "Indicates login success or failure."
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "id",
                        "name",
                        "email",
                        "isEmailVerified",
                        "account",
                        "role",
                        "createdAt",
                        "updatedAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        },
                        "name": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "example": "user@example.com"
                        },
                        "isEmailVerified": {
                          "type": "boolean",
                          "example": true
                        },
                        "account": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid",
                              "example": "987f6543-e21a-45bc-94d1-112233445566"
                            },
                            "name": {
                              "type": "string",
                              "example": "Example Corp"
                            }
                          }
                        },
                        "role": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid",
                              "example": "123e4567-e89b-12d3-a456-426614174000"
                            },
                            "code": {
                              "type": "string",
                              "example": "admin"
                            },
                            "value": {
                              "type": "string",
                              "example": "Administrator"
                            }
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2023-01-01T00:00:00.000Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2023-01-01T00:00:00.000Z"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2023-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Returned when the login request payload is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "errorCode",
                    "timestamp",
                    "path"
                  ],
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "FAILURE"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid login payload."
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "INVALID_LOGIN_CREDENTIALS"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2023-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/agent-backend/auth/login"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-email": {
      "get": {
        "operationId": "AuthController_verifyEmail",
        "summary": "Verify email address",
        "description": "Verifies the user's email using the token provided in the query parameter.",
        "tags": ["Auth"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Response after successful email verification.",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "SUCCESS"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "EMAIL_VERIFIED"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-07-25T07:38:36.424Z"
                    }
                  }
                },
                "example": {
                  "status": "SUCCESS",
                  "data": {
                    "message": "EMAIL_VERIFIED"
                  },
                  "timestamp": "2025-07-25T07:38:36.424Z"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired verification token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Failure response when email verification fails.",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "FAILURE"
                    },
                    "message": {
                      "type": "string",
                      "example": "User not found for the provided token"
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "HTTP_ERROR"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-07-25T07:38:36.424Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/agent-backend/auth/verify-email?token=..."
                    }
                  }
                },
                "example": {
                  "status": "FAILURE",
                  "message": "User not found for the provided token",
                  "errorCode": "HTTP_ERROR",
                  "timestamp": "2025-07-25T07:38:36.424Z",
                  "path": "/agent-backend/auth/verify-email?token=..."
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/invite": {
      "post": {
        "operationId": "AuthController_inviteUser",
        "summary": "Invite a new user",
        "description": "This endpoint is used to send an invitation to a user by email. It accepts user details and sends a verification email.",
        "tags": ["Auth"],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User invited successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Response after sending invitation(s) to user(s).",
                  "properties": {
                    "status": {
                      "type": "string",
                      "description": "Status of the request.",
                      "example": "SUCCESS"
                    },
                    "data": {
                      "type": "array",
                      "description": "List of user invitation responses.",
                      "items": {
                        "type": "object",
                        "properties": {
                          "email": {
                            "type": "string",
                            "format": "email",
                            "description": "Email address of the invited user.",
                            "example": "user@example.com"
                          },
                          "status": {
                            "type": "string",
                            "description": "Current status of the invitation process.",
                            "example": "completed"
                          },
                          "message": {
                            "type": "string",
                            "description": "Additional information about the invitation status.",
                            "example": "Verification email sent to user."
                          }
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Timestamp of the response generation.",
                      "example": "2025-07-25T07:15:53.996Z"
                    }
                  }
                },
                "example": {
                  "status": "SUCCESS",
                  "data": [
                    {
                      "email": "user@example.com",
                      "status": "completed",
                      "message": "Verification email sent to user."
                    }
                  ],
                  "timestamp": "2025-07-25T07:15:53.996Z"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Standard failure response structure for invalid inputs.",
                  "properties": {
                    "status": {
                      "type": "string",
                      "description": "Status of the request.",
                      "example": "FAILURE"
                    },
                    "message": {
                      "type": "string",
                      "description": "Optional human-readable error message.",
                      "example": ""
                    },
                    "errorCode": {
                      "type": "string",
                      "description": "Internal error code representing the failure reason.",
                      "example": "INVALID_INPUT"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Timestamp of the error.",
                      "example": "2025-07-25T07:25:58.025Z"
                    },
                    "path": {
                      "type": "string",
                      "description": "Request path that caused the error.",
                      "example": "/agent-backend/auth/invite"
                    }
                  }
                },
                "example": {
                  "status": "FAILURE",
                  "message": "",
                  "errorCode": "INVALID_INPUT",
                  "timestamp": "2025-07-25T07:25:58.025Z",
                  "path": "/agent-backend/auth/invite"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Agent Backend Service",
    "description": "✨The Agent Backend Service is a scalable, modular backend system designed to power intelligent agent-based applications. It provides robust APIs for authentication, user management, document processing, conversation tracking, training resources, and more. Built with flexibility in mind, it integrates seamlessly with external services while ensuring security, extensibility, and high availability.\n\n# 🔐 Authentication\n\nThis API uses a secure cookie named `token` to authenticate requests.\n\n- You must log in to receive this cookie.\n- When using **Try It Out** in the docs, **you must manually set the `token` cookie** in your browser.\n- You can find it in dev tools → Application → Cookies → {current url} -> `token`.\n\n**Note**: Scalar does not automatically send cookies during \"Try it out\".\n",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "name": "prod",
      "url": "https://zeno.sarvaha.ai/agent-backend"
    },
    {
      "name": "dev",
      "url": "http://localhost:7082/agent-backend"
    }
  ],
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com",
            "description": "User's email address used for login."
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "yourStrongPassword123",
            "description": "User's password used for login."
          }
        }
      },
      "InviteDto": {
        "type": "object",
        "properties": {
          "bulk": {
            "type": "boolean",
            "example": false,
            "description": "Indicates if the invite operation is for bulk users."
          },
          "userDetails": {
            "$ref": "#/components/schemas/SignupDto"
          }
        }
      },
...
      "SignupDto": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com",
            "description": "User's email address"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe",
            "description": "Full name of the user"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "superadmin", "guest", "user"],
            "example": "admin",
            "description": "User role (e.g., admin, user, guest)"
          },
          "accountName": {
            "type": "string",
            "example": "Acme Corp",
            "description": "Name of the associated account or organization"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 5,
            "example": "StrongPass123!",
            "description": "User's password (min. 5 characters)"
          }
        }
      }
    }
  }
}